Prevent 'receiveOneMsg' goroutine leak.
Per the receiveOneMsg function doc., the provided channel will be
written to at least one time. We did not buffer the provided
channel by that value. In an error scenario, this could result
in a leaked goroutine. I.e., it writes to a channel that has been
abandoned by the owner.